requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Dance[github]
python-dotenv
psycopg2-binary